\name{freq}
\alias{freq}
\docType{data}
\title{
A vector with note frequencies
}
\description{
A vector with note frequencies
}
\usage{data(freq)}
\format{
  The format is:
 Named num [1:96] 55 58.3 61.7 65.4 69.3 ...
 - attr(*, "names")= chr [1:96] "A1" "A#1" "B1" "C2" ...
}
\details{
}
\source{
MUSIC THEORY
}

\examples{
data(freq)
## maybe str(freq) ; plot(freq) ...
}
\keyword{datasets}
