\docType{data}
\name{furelise}
\alias{furelise}
\title{Fur Elise}
\format{A character vector of notes}
\description{
  The song "Fur Elise" by Beethoven.
}
\examples{
song <- makesong(furelise, baseduration = 1/3)
playsong(song)
}
\references{
  \url{http://en.wikipedia.org/wiki/F%C3%BCr_Elise}
}
\keyword{datasets}

