\docType{package}
\name{musicmakeR}
\alias{musicmakeR}
\alias{musicmakeR-package}
\alias{package-musicmakeR}
\title{Provides simple functions to make creating music in R easy.}
\description{
  This package provides an easy way to create music from
  within R. Currently only simple sine waves are used but
  chords can be created quite easily as well.
}

