\docType{data}
\name{odetojoy}
\alias{odetojoy}
\title{Ode to Joy}
\format{A character vector of notes}
\description{
  The song "Ode to Joy" by Beethoven.
}
\examples{
song <- makesong(odetojoy, baseduration = 1/4)
playsong(song)
}
\references{
  \url{http://en.wikipedia.org/wiki/Symphony_No._9_(Beethoven)}
}
\keyword{datasets}

